Move Command class tests to separate test file
